Move scope test recorder into test package
