+++ packages/cursorless-engine/src/index.ts
@@ -10,7 +10,6 @@ export * from "./CommandRunner";
 export * from "./CommandHistory";
 export * from "./CommandHistoryAnalyzer";
 export * from "./util/grammarHelpers";
-export * from "./ScopeTestRecorder";
 export * from "./core/Snippets";
 export * from "./core/mergeSnippets";
 export * from "./core/StoredTargets";
