+++ packages/cursorless-vscode/src/registerCommands.ts
@@ -11,9 +11,11 @@ import {
   CommandApi,
   StoredTargetMap,
   analyzeCommandHistory,
-  type ScopeTestRecorder,
 } from "@cursorless/cursorless-engine";
-import type { TestCaseRecorder } from "@cursorless/test-case-recorder";
+import type {
+  ScopeTestRecorder,
+  TestCaseRecorder,
+} from "@cursorless/test-case-recorder";
 import * as vscode from "vscode";
 import { ScopeVisualizer } from "./ScopeVisualizerCommandApi";
 import { showDocumentation, showQuickPick } from "./commands";
