+++ packages/test-case-recorder/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./TestCaseRecorder";
+export * from "./ScopeTestRecorder";
